Finished mongo datastore refactoring
